requirements.txt
python-qpid-proton
argparse
kafka
pyyaml
prometheus_client
jsonpath-rw
Flask
gunicorn
requests